services/order-service-django/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-redis==5.4.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
kafka-python==2.0.2
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.2.0
reportlab==4.0.9
django-filter==23.5
drf-yasg==1.21.7
sentry-sdk==1.40.0
prometheus-client==0.19.0
